+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { ActorPF2e, TokenPF2e, MeasuredTemplateDocumentPF2e } from "foundry-pf2e";
+import { ActorPF2e, TokenPF2e, MeasuredTemplateDocumentPF2e, ItemPF2e, ConditionPF2e, EffectPF2e } from "foundry-pf2e";
 import { getSetting, SETTINGS } from "./settings.ts";
 
 export const MODULE_ID = "samioli-module";
@@ -140,4 +140,35 @@ export async function deleteLightFromTemplate(template: MeasuredTemplateDocument
     } else {
         logd(`Light with ID: ${lightId} not found on the canvas.`);
     }
+}
+
+export function isCondition(item: ItemPF2e) : item is ConditionPF2e {  
+    return item.type === "condition";  
+}
+
+export function isEffect(item: ItemPF2e) : item is EffectPF2e {  
+    return item.type === "effect";  
+}
+
+export async function sendBasicChatMessage(content: string, recipients: string[], speaker: ActorPF2e) {
+    await ChatMessage.create({
+        content: content,
+        whisper: recipients,
+        speaker: ChatMessage.getSpeaker({ actor: speaker }),
+    });
+}
+
+export function returnStringOfNamesFromArray(names: string[]): string {
+  if (names.length === 0) {
+    return "";
+  }
+
+  if (names.length === 1) {
+    return names[0];
+  }
+
+  const allButLast = names.slice(0, -1).join(", ");
+  const last = names[names.length - 1];
+
+  return `${allButLast} and ${last}`;
 }